remove xxx.py and do clean ups
