+++ dataladmetadatamodel/mapper/__init__.py
@@ -1,44 +1,31 @@
+from _collections import defaultdict
+from typing import (
+    DefaultDict,
+    Dict
+)
 
+from .gitmapper import GIT_MAPPER_LOCATIONS
 from .memorymapper import (
     MEMORY_MAPPER_FAMILY,
     MEMORY_MAPPER_LOCATIONS
 )
-from .gitmapper import (
-    GIT_MAPPER_FAMILY_MEMBERS,
-    GIT_MAPPER_LOCATIONS
-)
 
 
-GIT_MAPPER_FAMILY_NAME = "git"
-MEMORY_MAPPER_FAMILY_NAME = "memory"
+git_mapper_family_name = "git"
+memory_mapper_family_name = "memory"
 
 
-MAPPER_FAMILIES = {
-    MEMORY_MAPPER_FAMILY_NAME: MEMORY_MAPPER_FAMILY,
-    GIT_MAPPER_FAMILY_NAME: GIT_MAPPER_FAMILY_MEMBERS
-}
-
-LOCATIONS = {
-    MEMORY_MAPPER_FAMILY_NAME: MEMORY_MAPPER_LOCATIONS,
-    GIT_MAPPER_FAMILY_NAME: GIT_MAPPER_LOCATIONS
+locations = {
+    memory_mapper_family_name: MEMORY_MAPPER_LOCATIONS,
+    git_mapper_family_name: GIT_MAPPER_LOCATIONS
 }
 
 
-def get_mapper(mapper_family: str, class_name: str):
-    family_class_mappers = MAPPER_FAMILIES.get(mapper_family, None)
-    if family_class_mappers is None:
-        raise ValueError(f"Unknown mapper family: {mapper_family}")
-
-    class_mapper = family_class_mappers.get(class_name, None)
-    if class_mapper is None:
-        raise ValueError(
-            f"No mapper for class: '{class_name}' "
-            f"in mapper family: '{mapper_family}'")
-    return class_mapper
+registered_mapper: DefaultDict[str, Dict[str, "Mapper"]] = defaultdict(dict)
 
 
 def _get_locations(mapper_family: str):
-    return LOCATIONS.get(mapper_family, (None, None))
+    return locations.get(mapper_family, (None, None))
 
 
 def get_uuid_set_location(mapper_family: str):
@@ -47,3 +34,36 @@ def get_uuid_set_location(mapper_family: str):
 
 def get_tree_version_list_location(mapper_family: str):
     return _get_locations(mapper_family)[0]
+
+
+def get_mapper(class_name: str, backend_type: str) -> "Mapper":
+    return registered_mapper[backend_type][class_name]
+
+
+def set_mapper(class_name: str, backend_type: str, mapper: "Mapper"):
+    registered_mapper[backend_type][class_name] = mapper
+
+
+def initialize_object_store():
+    from .gitmapper.datasettreemapper import DatasetTreeGitMapper
+    from .gitmapper.filetreemapper import FileTreeGitMapper
+    from .gitmapper.metadatamapper import MetadataGitMapper
+    from .gitmapper.metadatarootrecordmapper import MetadataRootRecordGitMapper
+    from .gitmapper.textmapper import TextGitMapper
+    from .gitmapper.uuidsetmapper import UUIDSetGitMapper
+    from .gitmapper.versionlistmapper import (
+        TreeVersionListGitMapper,
+        VersionListGitMapper
+    )
+
+    set_mapper("DatasetTree", "git", DatasetTreeGitMapper("DatasetTree"))
+    set_mapper("FileTree", "git", FileTreeGitMapper("FileTree"))
+    set_mapper("Metadata", "git", MetadataGitMapper("Metadata"))
+    set_mapper("MetadataRootRecord", "git", MetadataRootRecordGitMapper("MetadataRootRecord"))
+    set_mapper("Text", "git", TextGitMapper("Text"))
+    set_mapper("TreeVersionList", "git", TreeVersionListGitMapper("TreeVersionList"))
+    set_mapper("UUIDSet", "git", UUIDSetGitMapper("UUIDSet"))
+    set_mapper("VersionList", "git", VersionListGitMapper("VersionList"))
+
+
+initialize_object_store()
