+++ dataladmetadatamodel/versionlist.py
@@ -162,27 +162,24 @@ class TreeVersionList(VersionList):
 
     def deepcopy(self,
                  new_mapper_family: Optional[str] = None,
-                 new_realm: Optional[str] = None,
+                 new_destination: Optional[str] = None,
                  path_prefix: Optional[MetadataPath] = None
                  ) -> "TreeVersionList":
 
-        raise NotImplementedError
-        new_mapper_family = new_mapper_family or self.mapper_family
-        new_realm = new_realm or self.realm
         path_prefix = path_prefix or MetadataPath("")
 
-        copied_version_list = TreeVersionList(new_mapper_family, new_realm)
-
+        copied_version_list = TreeVersionList()
         for primary_data_version, version_record in self.version_set.items():
 
-            metadata_root_record = version_record.element_connector.load_object()
+            metadata_root_record: MetadataRootRecord = version_record.element.read_in()
 
             copied_version_list.set_versioned_element(
                 primary_data_version,
                 version_record.time_stamp,
                 path_prefix / version_record.path,
-                metadata_root_record.deepcopy(new_mapper_family, new_realm))
+                metadata_root_record.deepcopy(new_mapper_family, new_destination))
 
-            version_record.element_connector.purge()
+            version_record.element.write_out(new_destination)
+            version_record.element.purge()
 
         return copied_version_list
